refactor(customer-analytics): Move to canonical presentation layout and enable tach interface
Generated-By: PostHog Code
Task-Id: 2fcc28eb-d694-4a38-98e0-5956a3dceb15
