--- products/customer_analytics/backend/max_tools/__init__.py
@@ -1,5 +1,0 @@
-from products.customer_analytics.backend.max_tools.open_account import OpenAccountTool
-from products.customer_analytics.backend.max_tools.upsert_account import UpsertAccountTool
-from products.customer_analytics.backend.max_tools.upsert_account_notebook import UpsertAccountNotebookTool
-
-__all__ = ["OpenAccountTool", "UpsertAccountNotebookTool", "UpsertAccountTool"]
