+++ products/customer_analytics/backend/presentation/views/utils.py
