+++ products/customer_analytics/tsconfig.json
@@ -1,0 +1,6 @@
+{
+    "extends": "../../tsconfig.json",
+    "compilerOptions": {
+        "composite": true
+    }
+}
