+++ pyproject.toml
@@ -569,6 +569,16 @@ ignore_imports = [
     "products.logs.backend.presentation.views.api -> products.exports.backend.models.exported_asset",
     "products.logs.backend.presentation.views.alerts_api -> products.cdp.backend.api.hog_function",
     "products.logs.backend.presentation.views.alerts_api -> products.cdp.backend.models.hog_functions.hog_function",
+    # TODO: customer_analytics presentation wave — these views still reach internals directly.
+    # Thin each view to parse -> facade -> serialize and delete its entry. backend:contract-check
+    # stays off until this list is empty.
+    "products.customer_analytics.backend.presentation.views.external -> products.customer_analytics.backend.constants",
+    "products.customer_analytics.backend.presentation.views.external -> products.customer_analytics.backend.models",
+    "products.customer_analytics.backend.presentation.views.external -> products.customer_analytics.backend.models.account",
+    "products.customer_analytics.backend.presentation.views.organization_members -> products.customer_analytics.backend.constants",
+    "products.customer_analytics.backend.presentation.views.serializers -> products.customer_analytics.backend.models",
+    "products.customer_analytics.backend.presentation.views.utils -> products.customer_analytics.backend.models",
+    "products.customer_analytics.backend.presentation.views.views -> products.customer_analytics.backend.models",
 ]
 
 [tool.ty.rules]
